requirements.txt
asgiref==3.5.2
django-environ==0.9.0
django-filter==22.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
Markdown==3.4.1
pip==21.3.1
pytz==2022.5
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.5
wheel==0.37.1